+++ elite_dangerous_colonise/Models/Database Types/RingInsertType.cs
@@ -6,14 +6,14 @@ namespace elite_dangerous_colonise.Models.Database_Types
     public class RingInsertType
     {
         [PgName("systemID")]
-        public long SystemID { get; set; }
+        public decimal SystemID { get; set; }
         [PgName("ringName")]
         public string RingName { get; set; }
         [PgName("ringType")]
         public RingType Type { get; set; }
 
         public RingInsertType() { }
-        public RingInsertType(long systemID, string ringName, RingType ringType)
+        public RingInsertType(ulong systemID, string ringName, RingType ringType)
         {
             SystemID = systemID;
             RingName = ringName;
