+++ elite_dangerous_colonise/Models/Database Types/StationInsertType.cs
@@ -8,14 +8,14 @@ namespace elite_dangerous_colonise.Models.Database_Types
         [PgName("stationID")]
         public decimal StationID { get; set; }
         [PgName("systemID")]
-        public long SystemID { get; set; }
+        public decimal SystemID { get; set; }
         [PgName("stationName")]
         public string StationName {  get; set; }
         [PgName("controllingFaction")]
         public string ControllingFaction {  get; set; }
 
         public StationInsertType() { }
-        public StationInsertType(ulong stationID, long systemID, string stationName, string controllingFaction)
+        public StationInsertType(ulong stationID, ulong systemID, string stationName, string controllingFaction)
         {
             StationID = stationID;
             SystemID = systemID;
